Update to a newer Telegram API, needed by the end of the year. That updates userID ints to int64 and some other assorted small fixes.
Also add tags to struct fields so marshalling works.

+++ confighandler/config_handler.go
@@ -7,7 +7,7 @@ import (
 )
 
 type StreamjuryConfig struct {
-	SuperUserId    int
+	SuperUserId    int64
 	ChannelId      int64
 	ApiKey         string
 	ResultsAbsPath string
